Index

Initialize Parameters

SegregatedVMSSolver.init_params — Function

init_params(params::Dict{Symbol,Any})

It initialize the parametes of the simulations. Default values are added if not specified by the user. In case of a restarting simulation it reads the file and create a initial_rescale_factor<1 for CFL ramping initialization.

SegregatedVMSSolver.verifykey — Function

verifykey(params::Dict{Symbol,Any},keyname; val = false)

It check if the dictionary params has the entry keyname. If not it adds the new entry with the value val. It is used to add default values

Commons Procedures

SegregatedVMSSolver.print_model — Function

print_model(params::Dict{Symbol, Any})

It prints the model mesh

SegregatedVMSSolver.creation_fe_spaces — Function

creationfespaces(params::Dict{Symbol,Any}, udiritags, udirivalues, pdiritags, pdirivalues)

It creates the finite elements spaces accordingly to the previously generated dirichelet tags

SegregatedVMSSolver.create_initial_conditions — Function

createinitialconditions(params::Dict{Symbol,Any})

It creates the initial conditions for velocity and pressure. If restart is true then the velocity and the pressure field are interpoled on the specified DataFrame.

SegregatedVMSSolver.create_PETSc_setup — Function

createPETScsetup(M::AbstractMatrix,ksp_setup::Function)

Wrapper for creating PETSc symbolic and numeric setup for GridapPETSc

SegregatedVMSSolver.solve_case — Function

solve_case(params::Dict{Symbol,Any})

It solves iteratively the velocity and pressure system.

AddNewTags

SegregatedVMSSolver.create_new_tag! — Function

create_new_tag!(model::GridapDistributed.DistributedDiscreteModel, tagname::String, is_tag::Function)

It creates the centre tag at the tag_coordinate (Point); if mesh extremely fine the tolrances have to be smaller (unlikely)

SegregatedVMSSolver.add_new_tag! — Function

add_new_tag!(model, tag_coordinate::Point, tagname::String)

It add a new tag named tagname at the specified tag_coordinates

SegregatedVMSSolver.add_centre_tag! — Function

add_centre_tag!(model, tag_coordinate::Point)

It creates the centre tag at the tag_coordinate (Point); if mesh extremely fine the tolrances have to be smaller (unlikely)

Stabilization Parameters

SegregatedVMSSolver.h_param — Function

h_param(Ω::Triangulation, D::Int64)

For a given triangulation Ω it computes the cell size $Area^{1/2}, D=2$ $Volume^{1/3}, D=3$

SegregatedVMSSolver.G_params — Function

G_params(Ω::Triangulation, params)

Compute the tensor G and the values GG and gg according to the VMS formulation proposed by [5]

SegregatedVMSSolver.compute_d — Function

compute_d(trian::Gridap.Geometry.BodyFittedTriangulation, params) #trian == Ω

The inverse of the cell-map-field. It is evaluted in the middle of the refernce domain.

SegregatedVMSSolver.compute_G — Function

compute_G(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute G (AbstractArray of TensorValues)

SegregatedVMSSolver.compute_GG — Function

compute_GG(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute GG

SegregatedVMSSolver.compute_gg — Function

compute_GG(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute gg

Linear Utilities

SegregatedVMSSolver.create_ũ_vector — Function

createũvector(zfv1::AbstractVector)

It allocates the vector to keep in memory the velocity field up to previous 4 time steps

SegregatedVMSSolver.update_ũ_vector! — Function

updateũvector!(ũvec::Vector, uhnew::AbstractVector)

It updates the vector which stores the values of velocity at previous time steps.

SegregatedVMSSolver.update_ũ — Function

updateũ(ũvec::Vector)

It uses the Taylor expansion proposed by [6]

Stabilized Equations

SegregatedVMSSolver.cconv — Function

cconv(uadv, ∇u)

Wrapper for the convective term $u(\nabla u)$

SegregatedVMSSolver.segregated_equations_SUPG! — Function

segregatedequationsSUPG!(u_adv, params)

It provides the SUPG segregated and linearized equations following the model proposed by [2]

SegregatedVMSSolver.segregated_equations_VMS! — Function

segregatedequationsSUPG!(u_adv, params)

It provides the VMS segregated and linearized equations

SolversOptions

SegregatedVMSSolver.petsc_options — Function

petscoptions(; velksp="gmres", velpc="gamg", presksp = "cg", pres_pc = "gamg") It provides the command-line for GridapPETSc to solve the segregated linear systems

MatrixCreation

SegregatedVMSSolver.allocate_Mat_inv_ML — Function

allocateMatinvML(MatML::PSparseMatrix)

It allocates a zero vector where to store the inverse of the lumped matrix

SegregatedVMSSolver.inv_lump_vel_mass! — Function

invlumpvelmass!(MatinvML::PVector,MatML::PSparseMatrix)

It computes the lumped matrix, takes the inverse of the diagonal elements.

SegregatedVMSSolver.initialize_vectors — Function

initialize_vectors(matrices::Tuple,uh0,ph0)

It initializes vectors where velocity, pressure, acceleration and all the increments will be stored.

SegregatedVMSSolver.matrices_and_vectors — Function

matricesandvectors(trials, tests, t::Real, u_adv, params; method=:VMS)

It updates matrices and vectors

Restart

SegregatedVMSSolver.find_idx — Function

find_idx(p::VectorValue{3, Float64}, params)

For a given point p it search on the restart_file provided by the user the closer node and provides the index of such node.

SegregatedVMSSolver.uh_restart — Function

uh_restart(p, params::Dict{Symbol, Any})

For a given point p it calles find_idx which provide the line of the csv file corresponding to that point. Then, it calles uh which provide the VectorValue of the velocity at that point.

SegregatedVMSSolver.ph_restart — Function

ph_restart(p, params::Dict{Symbol, Any})

For a given point p it calles find_idx which provide the line of the csv file corresponding to that point. Then, it calles ph which provide the scalar pressure value at that point.

SegregatedVMSSolver.restart_uh_field — Function

restart_uh_field(params::Dict{Symbol, Any})

It provides a suitable function which gives for each point the specified velocity in restart_file. It is used as initial condition for restarting a simulation at a specific time step.

SegregatedVMSSolver.restart_ph_field — Function

restart_ph_field(params::Dict{Symbol, Any})

It provides a suitable function which gives for each point the specified pressure in restart_file. It is used as initial condition for restarting a simulation at a specific time step.

SegregatedVMSSolver.ExportUtility

SegregatedVMSSolver.ExportUtility.conv_to_df — Method

conv_to_df(vv::Vector)

Convert a Vector{Float64} to a DataFrame. It is used for export pressure field.

SegregatedVMSSolver.ExportUtility.conv_to_df — Method

conv_to_df(vv::Vector)

Convert a Vector{Vector} to a DataFrame. It is used for export nodes, normals.

SegregatedVMSSolver.ExportUtility.create_export_tags! — Method

create_export_tags!(params::Dict{Symbol,Any})

For each name_tags specified, it creates the boundary triangulation and the normal vector fields

SegregatedVMSSolver.ExportUtility.export_fields — Method

export_fields(params::Dict{Symbol,Any}, local_unique_idx, global_unique_idx, tt::Float64, uh0, ph0)

Export pressure and friction (not multiplied by viscosity) - airfoil simulations oriented

SegregatedVMSSolver.ExportUtility.print_on_request — Method

print_on_request(log_dir::String)

If in the directory log_dir exists and there is the PrintSim.txt file return true. If true is printing simulations results in Paraview format (.vtu and .pvtu).

SegregatedVMSSolver.ReadAirfoilResults

SegregatedVMSSolver.ReadAirfoilResults.average_field — Method

average_field(path::String, field_name::String, nodes::DataFrame; offset=1, offend_ = 0)

It computes the time-average and also the spanwise averge (z direction). field_name can be ph for pressure or friction for friction. It is possible to skip a certain amount of initial time-step when averaging setting the offset.

SegregatedVMSSolver.ReadAirfoilResults.compute_CL_CD — Method

compute_CL_CD(top_nodesx,bottom_nodesx,top_nodesy,bottom_nodesy,cp_top,cp_bottom,
friction_top,friction_bottom; chord = 1.0)

It computes lift and drag coefficients

SegregatedVMSSolver.ReadAirfoilResults.extract_airfoil_features — Method

extract_airfoil_features(nodes::DataFrame, n_Γ0::DataFrame, Ph::DataFrame, Friction::DataFrame; u0::Float64, μ::Float64, rho::Float64, α::Float64, chord::Float64)

It provides the airfoil features like pressure and friction coefficient splitted between top and bottom. The nodes and results are orderded for growing x coordinates.

SegregatedVMSSolver.ReadAirfoilResults.get_nodes — Method

get_nodes(path::String)

It provides a DataFrame with the nodes of the Airfoil boundary

SegregatedVMSSolver.ReadAirfoilResults.get_normals — Method

get_nodes(path::String)

It provides a DataFrame with the normals vectors at the surface of the Airfoil boundary

SegregatedVMSSolver.WallDistance

SegregatedVMSSolver.WallDistance.get_initial_conditions — Method

get_initial_conditions(mesh_file::String, D::Int64, u_in::Float64, Re::Real, chord::Float64, walltag)

It creates the .vtu file with the boundary layer initialization and also the .csv file which can be used to start a simulation over an airfoil. It return a dataframe where for each nodes are provided the coordinates and the velocity vector components.